+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialWorldSettings.h
@@ -6,6 +6,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/WorldSettings.h"
+#include "SpatialGDKSettings.h"
 #include "Utils/LayerInfo.h"
 
 #include "SpatialWorldSettings.generated.h"
@@ -32,4 +33,14 @@ public:
 	/** Layer configuration. */
 	UPROPERTY(EditAnywhere, Config, Category = "Multi-Worker", meta = (EditCondition = "bEnableMultiWorker"))
 	TMap<FName, FLayerInfo> WorkerLayers;
+
+	bool IsMultiWorkerEnabled() const
+	{
+		const USpatialGDKSettings* SpatialGDKSettings = GetDefault<USpatialGDKSettings>();
+		if (SpatialGDKSettings->bOverrideMultiWorker.IsSet())
+		{
+			return SpatialGDKSettings->bOverrideMultiWorker.GetValue();
+		}
+		return bEnableMultiWorker;
+	}
 };